A lot of changes.
First of all, now it is using String type instead of Enum. Strings is easier to read in serialized data, easy to change without keeping in mind ids, do not require recompile on change, etc. Also to make it easier to use strings, there is Consts code generator now, which is makes static class with consts, similar to Enum.
Structure changes.
Example was updated and improved a bit. More comments were added.